Also remove `Node` class
So just have the single class which composes itself.
