+++ file-backup-core/src/main/java/com/willmolloy/backup/FileTree.java
@@ -62,7 +62,7 @@ public interface FileTree<FileT extends File> {
    */
   static <FileT extends File> Builder<FileT> builder(
       FileT root, DirectoryFiller<FileT> directoryFiller) {
-    return new TrieLikeFileTree.Builder<>(root, directoryFiller);
+    return new FileTreeNode.Builder<>(root, directoryFiller);
   }
 
   /**
