+++ file-backup-core/src/main/java/com/willmolloy/backup/FileTreeNode.java
@@ -1,0 +1,182 @@
+package com.willmolloy.backup;
+
+import static com.willmolloy.backup.util.PathHelper.nameComponents;
+import static java.util.Objects.requireNonNull;
+import static java.util.function.Predicate.not;
+
+import java.nio.file.Path;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Stream;
+import javax.annotation.Nullable;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+/**
+ * {@link FileTree} implemented like an n-ary tree/trie over {@link Path} {@linkplain Path#iterator
+ * name components}.
+ *
+ * @param <FileT> type of file stored in this file tree
+ * @author <a href=https://willmolloy.com>Will Molloy</a>
+ */
+final class FileTreeNode<FileT extends File> implements FileTree<FileT> {
+
+  private static final Logger log = LogManager.getLogger();
+
+  private final FileT file;
+  @Nullable private final FileTreeNode<FileT> parent;
+  private final Map<String, FileTreeNode<FileT>> children;
+
+  private FileTreeNode(FileT file, FileTreeNode<FileT> parent) {
+    this.file = requireNonNull(file);
+    this.parent = parent;
+    this.children = new LinkedHashMap<>();
+  }
+
+  @Override
+  public Optional<FileT> get(Path relativePath) {
+    return getNode(relativePath).map(getFile);
+  }
+
+  @Override
+  public boolean isRoot(FileT file) {
+    return file == this.file;
+  }
+
+  @Override
+  public Stream<FileT> postorder() {
+    return postorderNodes().map(getFile);
+  }
+
+  @Override
+  public Stream<FileT> leaves() {
+    return postorderNodes().filter(node -> node.children.isEmpty()).map(getFile);
+  }
+
+  @Override
+  public Stream<FileT> ancestors(FileT file) {
+    return getNode(file.relativePath()).stream()
+        .flatMap(node -> Stream.iterate(node.parent, parent -> parent.parent))
+        .takeWhile(node -> node != this.parent)
+        .map(getFile);
+  }
+
+  @Override
+  public FileTree<FileT> subtree(FileT file) {
+    return getNode(file.relativePath())
+        // shouldn't get here... only if you did subtree.subtree
+        .orElseThrow(IllegalArgumentException::new);
+  }
+
+  @Override
+  public long fileCount() {
+    return files().count();
+  }
+
+  @Override
+  public long totalSize() {
+    return files().mapToLong(File::size).sum();
+  }
+
+  private Stream<FileT> files() {
+    return postorder().filter(not(File::isDirectory));
+  }
+
+  private Optional<FileTreeNode<FileT>> getNode(Path path) {
+    FileTreeNode<FileT> node = this;
+    List<String> pathToNode = nameComponents(this.file.relativePath().relativize(path));
+    for (String c : pathToNode) {
+      FileTreeNode<FileT> child = node.children.get(c);
+      if (child == null) {
+        return Optional.empty();
+      }
+      node = child;
+    }
+    return Optional.of(node);
+  }
+
+  private void insertNode(FileT file, DirectoryFiller<FileT> directoryFiller) {
+    FileTreeNode<FileT> node = this;
+    StringBuilder pathSoFar = new StringBuilder(this.file.relativePath().toString());
+    List<String> pathToNode =
+        nameComponents(this.file.relativePath().relativize(file.relativePath()));
+    for (int i = 0; i < pathToNode.size(); i++) {
+      pathSoFar.append(pathToNode.get(i));
+      FileTreeNode<FileT> currentNode = node;
+      boolean last = i == pathToNode.size() - 1;
+      node =
+          node.children.computeIfAbsent(
+              pathToNode.get(i),
+              k ->
+                  last
+                      ? new FileTreeNode<>(file, currentNode)
+                      : new FileTreeNode<>(
+                          directoryFiller.apply(pathSoFar.toString()), currentNode));
+      pathSoFar.append('/');
+    }
+    // no need to set Node.file here; assuming only leaves are inserted or parent dirs are
+    // inserted first (i.e. in a pre-order manner)
+  }
+
+  private Stream<FileTreeNode<FileT>> postorderNodes() {
+    return Stream.concat(
+        children.values().stream().flatMap(FileTreeNode::postorderNodes), Stream.of(this));
+  }
+
+  private final Function<FileTreeNode<FileT>, FileT> getFile = node -> node.file;
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null) {
+      return false;
+    }
+    if (o instanceof FileTreeNode<?> node) {
+      return Objects.equals(file, node.file) && Objects.equals(children, node.children);
+    }
+    return false;
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(file, children);
+  }
+
+  @Override
+  public String toString() {
+    return "Node[file=%s, children=%s]".formatted(file, children);
+  }
+
+  /**
+   * {@link FileTree.Builder} implementation for {@link FileTreeNode}.
+   *
+   * @param <FileT> type of file stored in the built file tree
+   */
+  static final class Builder<FileT extends File> implements FileTree.Builder<FileT> {
+    private final FileTreeNode<FileT> root;
+    private final DirectoryFiller<FileT> directoryFiller;
+
+    Builder(FileT root, DirectoryFiller<FileT> directoryFiller) {
+      this.root = new FileTreeNode<>(root, null);
+      this.directoryFiller = requireNonNull(directoryFiller);
+    }
+
+    @Override
+    public Builder<FileT> insert(FileT file) {
+      log.debug("insert({})", file);
+      root.insertNode(file, directoryFiller);
+      return this;
+    }
+
+    @Override
+    public FileTreeNode<FileT> build() {
+      return root;
+    }
+  }
+}
